+++ lib/blame-status-bar-view.js
@@ -4,7 +4,7 @@ import gravatar from 'gravatar';
 import open from 'open';
 import moment from 'moment';
 import { CompositeDisposable, Disposable } from 'atom';
-import { isCommitted, findRepo, blame, getCommit, getCommitLink } from './utils';
+import { isCommitted, findRepo, blameFile, getCommit, getCommitLink } from './utils';
 
 function formatLine(hash, line) {
   const dateFormat = atom.config.get('status-bar-blame.dateFormat');
@@ -144,7 +144,7 @@ class BlameStatusBarView extends HTMLElement {
     const filePath = this.editor.getPath();
     if (!filePath) { return null; }
 
-    const result = await blame(this.editor.getPath());
+    const result = await blameFile(this.editor.getPath());
     if (!result) { return null; }
 
     return result.map((line) => {
