+++ lib/utils/index.js
@@ -1,8 +1,8 @@
 /** @babel */
 
-export blame from './blame';
+export * as git from './git';
+export blameFile from './blame-file';
 export findRepo from './find-repo';
 export getCommitLink from './get-commit-link';
 export getCommit from './get-commit';
 export isCommitted from './is-committed';
-export * as git from './git';
